SCP-2906: Deal with real Alonzo transactions in the wallet effect.
* Defined the type `type CardanoTx = Either SomeCardanoApiTx Tx`

* Changed anything that uses the wallet effect from Tx to CardanoTx.

* Moved functions from Plutus.Contract.CardanoAPI (plutus-chain-index) to Ledger.Tx.CardanoAPI (plutus-ledger).
