+++ plutus-ledger/src/Ledger/Tx.hs
@@ -7,6 +7,7 @@
 {-# LANGUAGE TypeApplications  #-}
 
 {-# OPTIONS_GHC -fno-warn-orphans #-}
+{-# LANGUAGE RecordWildCards   #-}
 
 module Ledger.Tx
     ( module Export
@@ -21,6 +22,11 @@ module Ledger.Tx
     , ciTxOutValidator
     , _PublicKeyChainIndexTxOut
     , _ScriptChainIndexTxOut
+    , CardanoTx
+    , getCardanoTxId
+    , getCardanoTxInputs
+    , getCardanoTxOutRefs
+    , getCardanoTxUnspentOutputsTx
     , SomeCardanoApiTx(..)
     -- * Transactions
     , addSignature
@@ -38,22 +44,20 @@ import qualified Cardano.Api               as C
 import           Cardano.Crypto.Hash       (SHA256, digest)
 import qualified Codec.CBOR.Write          as Write
 import           Codec.Serialise           (Serialise (..))
-import           Codec.Serialise.Decoding  (Decoder, decodeBytes, decodeSimple)
-import           Codec.Serialise.Encoding  (Encoding (..), Tokens (..))
-import           Control.Applicative       ((<|>))
 import           Control.Lens              hiding ((.=))
-import           Data.Aeson                (FromJSON (parseJSON), ToJSON (toJSON), object, (.:), (.=))
-import qualified Data.Aeson                as Aeson
-import           Data.Aeson.Types          (Parser, parseFail, prependFailure, typeMismatch)
+import           Data.Aeson                (FromJSON, ToJSON)
 import           Data.Map                  (Map)
 import qualified Data.Map                  as Map
 import           Data.Proxy
+import           Data.Set                  (Set)
 import qualified Data.Set                  as Set
 import           Data.Text.Prettyprint.Doc (Pretty (pretty), braces, colon, hang, nest, viaShow, vsep, (<+>))
 import           GHC.Generics              (Generic)
 import           Ledger.Address            (pubKeyAddress, scriptAddress)
 import           Ledger.Crypto             (PrivateKey, PubKey, signTx, toPublicKey)
 import           Ledger.Scripts            (datumHash)
+import           Ledger.Tx.CardanoAPI      (SomeCardanoApiTx (SomeTx))
+import qualified Ledger.Tx.CardanoAPI      as CardanoAPI
 import           Plutus.V1.Ledger.Api      (Credential (PubKeyCredential, ScriptCredential), Datum, DatumHash,
                                             TxId (..), Validator, ValidatorHash, Value, addressCredential, toBuiltin)
 import           Plutus.V1.Ledger.Tx       as Export
@@ -61,7 +65,7 @@ import           Plutus.V1.Ledger.Tx       as Export
 -- | Transaction output that comes from a chain index query.
 --
 -- It is defined here instead of the plutus-chain-index because plutus-ledger
--- uses that datatype, and plutus-chain-index can't depend on plutus-ledger
+-- uses that datatype, and plutus-ledger can't depend on plutus-chain-index
 -- because of a cyclic dependency.
 --
 -- This datatype was created in order to be used in
@@ -107,108 +111,24 @@ instance Pretty ChainIndexTxOut where
     pretty ScriptChainIndexTxOut {_ciTxOutAddress, _ciTxOutValue} =
                 hang 2 $ vsep ["-" <+> pretty _ciTxOutValue <+> "addressed to", pretty _ciTxOutAddress]
 
-data SomeCardanoApiTx where
-  SomeTx :: C.IsCardanoEra era => C.Tx era -> C.EraInMode era C.CardanoMode -> SomeCardanoApiTx
+type CardanoTx = Either SomeCardanoApiTx Tx
 
-instance Eq SomeCardanoApiTx where
-  (SomeTx tx1 C.ByronEraInCardanoMode) == (SomeTx tx2 C.ByronEraInCardanoMode)     = tx1 == tx2
-  (SomeTx tx1 C.ShelleyEraInCardanoMode) == (SomeTx tx2 C.ShelleyEraInCardanoMode) = tx1 == tx2
-  (SomeTx tx1 C.AllegraEraInCardanoMode) == (SomeTx tx2 C.AllegraEraInCardanoMode) = tx1 == tx2
-  (SomeTx tx1 C.MaryEraInCardanoMode) == (SomeTx tx2 C.MaryEraInCardanoMode)       = tx1 == tx2
-  (SomeTx tx1 C.AlonzoEraInCardanoMode) == (SomeTx tx2 C.AlonzoEraInCardanoMode)   = tx1 == tx2
-  _ == _                                                                           = False
+getCardanoTxId :: CardanoTx -> TxId
+getCardanoTxId (Left (SomeTx (C.Tx body _) _)) = CardanoAPI.fromCardanoTxId $ C.getTxId body
+getCardanoTxId (Right tx)                      = txId tx
 
-deriving instance Show SomeCardanoApiTx
+getCardanoTxInputs :: CardanoTx -> Set TxIn
+getCardanoTxInputs (Left (SomeTx (C.Tx (C.TxBody C.TxBodyContent {..}) _) _)) =
+  Set.fromList $ fmap ((`TxIn` Nothing) . CardanoAPI.fromCardanoTxIn . fst) txIns
+getCardanoTxInputs (Right tx) = txInputs tx
 
-instance Serialise SomeCardanoApiTx where
-  encode (SomeTx tx eraInMode) = encodedMode eraInMode <> Encoding (TkBytes (C.serialiseToCBOR tx))
-    where
-      encodedMode :: C.EraInMode era C.CardanoMode -> Encoding
-      -- 0 and 1 are for ByronEraInByronMode and ShelleyEraInShelleyMode
-      encodedMode C.ByronEraInCardanoMode   = Encoding (TkSimple 2)
-      encodedMode C.ShelleyEraInCardanoMode = Encoding (TkSimple 3)
-      encodedMode C.AllegraEraInCardanoMode = Encoding (TkSimple 4)
-      encodedMode C.MaryEraInCardanoMode    = Encoding (TkSimple 5)
-      encodedMode C.AlonzoEraInCardanoMode  = Encoding (TkSimple 6)
-  decode = do
-    w <- decodeSimple
-    case w of
-      2 -> decodeTx C.AsByronEra C.ByronEraInCardanoMode
-      3 -> decodeTx C.AsShelleyEra C.ShelleyEraInCardanoMode
-      4 -> decodeTx C.AsAllegraEra C.AllegraEraInCardanoMode
-      5 -> decodeTx C.AsMaryEra C.MaryEraInCardanoMode
-      6 -> decodeTx C.AsAlonzoEra C.AlonzoEraInCardanoMode
-      _ -> fail "Unexpected value while decoding Cardano.Api.EraInMode"
-    where
-      decodeTx :: C.IsCardanoEra era => C.AsType era -> C.EraInMode era C.CardanoMode -> Decoder s SomeCardanoApiTx
-      decodeTx asType eraInMode = do
-        bytes <- decodeBytes
-        tx <- either (const $ fail "Failed to decode Cardano.Api.Tx") pure $ C.deserialiseFromCBOR (C.AsTx asType) bytes
-        pure $ SomeTx tx eraInMode
+getCardanoTxOutRefs :: CardanoTx -> [(TxOut, TxOutRef)]
+getCardanoTxOutRefs (Left tx)  = CardanoAPI.txOutRefs tx
+getCardanoTxOutRefs (Right tx) = txOutRefs tx
 
-instance ToJSON SomeCardanoApiTx where
-  toJSON (SomeTx tx eraInMode) =
-    object [ "tx" .= C.serialiseToTextEnvelope Nothing tx
-           , "eraInMode" .= eraInMode
-           ]
-
--- | Converting 'SomeCardanoApiTx' to JSON.
---
--- If the "tx" field is from an unknown era, the JSON parser will print an
--- error at runtime while parsing.
-instance FromJSON SomeCardanoApiTx where
-  parseJSON v = parseByronInCardanoModeTx v
-            <|> parseShelleyEraInCardanoModeTx v
-            <|> parseAllegraEraInCardanoModeTx v
-            <|> parseMaryEraInCardanoModeTx v
-            <|> parseAlonzoEraInCardanoModeTx v
-            <|> parseEraInCardanoModeFail v
-
-parseByronInCardanoModeTx :: Aeson.Value -> Parser SomeCardanoApiTx
-parseByronInCardanoModeTx =
-  parseSomeCardanoTx "Failed to parse ByronEra 'tx' field from SomeCardanoApiTx"
-                     (C.AsTx C.AsByronEra)
-
-parseShelleyEraInCardanoModeTx :: Aeson.Value -> Parser SomeCardanoApiTx
-parseShelleyEraInCardanoModeTx =
-  parseSomeCardanoTx "Failed to parse ShelleyEra 'tx' field from SomeCardanoApiTx"
-                     (C.AsTx C.AsShelleyEra)
-
-parseMaryEraInCardanoModeTx :: Aeson.Value -> Parser SomeCardanoApiTx
-parseMaryEraInCardanoModeTx =
-  parseSomeCardanoTx "Failed to parse MaryEra 'tx' field from SomeCardanoApiTx"
-                     (C.AsTx C.AsMaryEra)
-
-parseAllegraEraInCardanoModeTx :: Aeson.Value -> Parser SomeCardanoApiTx
-parseAllegraEraInCardanoModeTx =
-  parseSomeCardanoTx "Failed to parse AllegraEra 'tx' field from SomeCardanoApiTx"
-                     (C.AsTx C.AsAllegraEra)
-
-parseAlonzoEraInCardanoModeTx :: Aeson.Value -> Parser SomeCardanoApiTx
-parseAlonzoEraInCardanoModeTx =
-  parseSomeCardanoTx "Failed to parse AlonzoEra 'tx' field from SomeCardanoApiTx"
-                     (C.AsTx C.AsAlonzoEra)
-
-parseEraInCardanoModeFail :: Aeson.Value -> Parser SomeCardanoApiTx
-parseEraInCardanoModeFail _ = fail "Unable to parse 'eraInMode'"
-
-parseSomeCardanoTx
-  :: ( FromJSON (C.EraInMode era C.CardanoMode)
-     , C.IsCardanoEra era
-     )
-  => String
-  -> C.AsType (C.Tx era)
-  -> Aeson.Value
-  -> Parser SomeCardanoApiTx
-parseSomeCardanoTx errorMsg txAsType (Aeson.Object v) =
-  SomeTx
-    <$> (v .: "tx" >>= \envelope -> either (const $ parseFail errorMsg)
-                                           pure
-                                           $ C.deserialiseFromTextEnvelope txAsType envelope)
-    <*> v .: "eraInMode"
-parseSomeCardanoTx _ _ invalid =
-    prependFailure "parsing SomeCardanoApiTx failed, "
-      (typeMismatch "Object" invalid)
+getCardanoTxUnspentOutputsTx :: CardanoTx -> Map TxOutRef TxOut
+getCardanoTxUnspentOutputsTx (Left tx)  = CardanoAPI.unspentOutputsTx tx
+getCardanoTxUnspentOutputsTx (Right tx) = unspentOutputsTx tx
 
 instance Pretty Tx where
     pretty t@Tx{txInputs, txCollateral, txOutputs, txMint, txFee, txValidRange, txSignatures, txMintScripts, txData} =
