+++ plutus-ledger/src/Ledger/Tx/CardanoAPITemp.hs
@@ -4,10 +4,9 @@
 {-# LANGUAGE FlexibleContexts         #-}
 {-# LANGUAGE GADTs                    #-}
 {-# LANGUAGE NamedFieldPuns           #-}
-{-# LANGUAGE TypeApplications         #-}
 -- Code temporarily copied over from cardano-api,
 -- until https://github.com/input-output-hk/cardano-node/pull/2936 or something similar gets merged.
-module Plutus.Contract.CardanoAPITemp (makeTransactionBody') where
+module Ledger.Tx.CardanoAPITemp (makeTransactionBody') where
 
 import qualified Data.Map.Strict                    as Map
 import qualified Data.Sequence.Strict               as Seq
@@ -157,3 +156,4 @@ toAlonzoTxOutDataHash :: TxOutDatumHash era
                       -> StrictMaybe (Alonzo.DataHash StandardCrypto)
 toAlonzoTxOutDataHash TxOutDatumHashNone                     = SNothing
 toAlonzoTxOutDataHash (TxOutDatumHash _ (ScriptDataHash dh)) = SJust dh
+
