+++ plutus-pab/plutus-pab.cabal
@@ -137,34 +137,20 @@ library
         base >=4.7 && <5,
         bytestring -any,
         cardano-api -any,
-        cardano-crypto -any,
-        cardano-ledger-byron -any,
-        cardano-prelude -any,
         cardano-slotting -any,
         containers -any,
         contra-tracer -any,
         cryptonite -any,
-        cborg -any,
         freer-simple -any,
-        fingertree -any,
         generic-arbitrary -any,
         http-client -any,
         http-client-tls -any,
-        http-types -any,
-        io-classes -any,
         lens -any,
         memory -any,
         monad-logger -any,
         mtl -any,
-        network -any,
         network-mux -any,
-        newtype-generics -any,
-        nothunks -any,
         optparse-applicative -any,
-        ouroboros-consensus -any,
-        ouroboros-consensus-byron -any,
-        ouroboros-consensus-shelley -any,
-        ouroboros-consensus-cardano -any,
         ouroboros-network -any,
         ouroboros-network-framework -any,
         playground-common -any,
@@ -172,10 +158,7 @@ library
         purescript-bridge -any,
         prettyprinter >=1.1.0.1,
         quickcheck-instances -any,
-        random -any,
         row-types -any,
-        scientific -any,
-        serialise -any,
         servant -any,
         servant-client -any,
         servant-server -any,
@@ -184,26 +167,19 @@ library
         openapi3 -any,
         servant-options -any,
         tagged -any,
-        typed-protocols -any,
-        typed-protocols-examples -any,
         servant-purescript -any,
         servant-websockets -any,
         stm -any,
         text -any,
         transformers -any,
-        unliftio-core -any,
         unordered-containers -any,
         uuid -any,
-        vector -any,
         warp -any,
         wai -any,
         wai-cors -any,
         Win32-network -any,
         websockets -any,
         yaml -any,
-        mwc-random -any,
-        primitive -any,
-        hedgehog -any,
         iohk-monitoring -any,
         lobemo-backend-ekg -any,
         exceptions -any,
