+++ plutus-pab/test/full/Plutus/PAB/CoreSpec.hs
@@ -43,8 +43,9 @@ import qualified Data.Set                                 as Set
 import           Data.Text                                (Text)
 import qualified Data.Text                                as Text
 import           Data.Text.Extras                         (tshow)
-import           Ledger                                   (pubKeyAddress, pubKeyHash, pubKeyHashAddress, toPubKeyHash,
-                                                           txId, txOutAddress, txOutRefId, txOutRefs, txOutputs)
+import           Ledger                                   (getCardanoTxId, getCardanoTxOutRefs, pubKeyAddress,
+                                                           pubKeyHash, pubKeyHashAddress, toPubKeyHash, txId,
+                                                           txOutAddress, txOutRefId, txOutRefs, txOutputs)
 import           Ledger.Ada                               (adaSymbol, adaToken, lovelaceValueOf)
 import qualified Ledger.Ada                               as Ada
 import qualified Ledger.AddressMap                        as AM
@@ -178,7 +179,7 @@ waitForTxStatusChangeTest = runScenario $ do
   (w1, pk1) <- Simulator.addWallet
   Simulator.waitNSlots 1
   tx <- Simulator.payToPublicKeyHash w1 pk1 (lovelaceValueOf 100_000_000)
-  txStatus <- Simulator.waitForTxStatusChange (txId tx)
+  txStatus <- Simulator.waitForTxStatusChange (getCardanoTxId tx)
   assertEqual "tx should be tentatively confirmed of depth 1"
               (TentativelyConfirmed 1 TxValid ())
               txStatus
@@ -187,7 +188,7 @@ waitForTxStatusChangeTest = runScenario $ do
   -- increment the block number.
   void $ Simulator.payToPublicKeyHash w1 pk1 (lovelaceValueOf 1_000_000)
   Simulator.waitNSlots 1
-  txStatus' <- Simulator.waitForTxStatusChange (txId tx)
+  txStatus' <- Simulator.waitForTxStatusChange (getCardanoTxId tx)
   assertEqual "tx should be tentatively confirmed of depth 2"
               (TentativelyConfirmed 2 TxValid ())
               txStatus'
@@ -198,7 +199,7 @@ waitForTxStatusChangeTest = runScenario $ do
     void $ Simulator.payToPublicKeyHash w1 pk1 (lovelaceValueOf 1_000_000)
     Simulator.waitNSlots 1
 
-  txStatus'' <- Simulator.waitForTxStatusChange (txId tx)
+  txStatus'' <- Simulator.waitForTxStatusChange (getCardanoTxId tx)
   assertEqual "tx should be committed"
               (Committed TxValid ())
               txStatus''
@@ -216,9 +217,9 @@ waitForTxOutStatusChangeTest = runScenario $ do
   tx <- Simulator.payToPublicKeyHash w1 pk2 (lovelaceValueOf 100_000_000)
   -- We should have 2 UTxOs present.
   -- We find the 'TxOutRef' from wallet 1
-  let txOutRef1 = head $ fmap snd $ filter (\(txOut, txOutref) -> toPubKeyHash (txOutAddress txOut) == Just pk1) $ txOutRefs tx
+  let txOutRef1 = head $ fmap snd $ filter (\(txOut, txOutref) -> toPubKeyHash (txOutAddress txOut) == Just pk1) $ getCardanoTxOutRefs tx
   -- We find the 'TxOutRef' from wallet 2
-  let txOutRef2 = head $ fmap snd $ filter (\(txOut, txOutref) -> toPubKeyHash (txOutAddress txOut) == Just pk2) $ txOutRefs tx
+  let txOutRef2 = head $ fmap snd $ filter (\(txOut, txOutref) -> toPubKeyHash (txOutAddress txOut) == Just pk2) $ getCardanoTxOutRefs tx
   txOutStatus1 <- Simulator.waitForTxOutStatusChange txOutRef1
   assertEqual "tx output 1 should be tentatively confirmed of depth 1"
               (TentativelyConfirmed 1 TxValid Unspent)
@@ -234,7 +235,7 @@ waitForTxOutStatusChangeTest = runScenario $ do
   Simulator.waitNSlots 1
   txOutStatus1' <- Simulator.waitForTxOutStatusChange txOutRef1
   assertEqual "tx output 1 should be tentatively confirmed of depth 1"
-              (TentativelyConfirmed 1 TxValid (Spent $ txId tx2))
+              (TentativelyConfirmed 1 TxValid (Spent $ getCardanoTxId tx2))
               txOutStatus1'
   txOutStatus2' <- Simulator.waitForTxOutStatusChange txOutRef2
   assertEqual "tx output 2 should be tentatively confirmed of depth 2"
@@ -249,7 +250,7 @@ waitForTxOutStatusChangeTest = runScenario $ do
 
   txOutStatus1'' <- Simulator.waitForTxOutStatusChange txOutRef1
   assertEqual "tx output 1 should be committed"
-              (Committed TxValid (Spent $ txId tx2))
+              (Committed TxValid (Spent $ getCardanoTxId tx2))
               txOutStatus1''
   txOutStatus2'' <- Simulator.waitForTxOutStatusChange txOutRef2
   assertEqual "tx output 2 should be committed"
